Extract help commands to seperate handler
